Error fixes and buzzer code
Today I fixed some errors after doing testing using the actual circuit. Also, I wanted to add buzzer code on my python script in pycharm instead of just having it on the Arduino IDE. All though this code doesn't actually control the buzzer, it should show a text in the middle of the graph saying "Buzzer On" once the thresholds are breached. It basically imitates the actual buzzer code in C++ on the Arduino IDE.